[FE] Refactor: toast 호출 방식 수정
