+++ frontend/src/components/common/Toast.tsx
@@ -1,5 +1,32 @@
-import { useToastStore } from "../../store/useToast";
 import { AnimatePresence, motion } from "framer-motion";
+import { ReactNode } from "react";
+import { create } from "zustand";
+
+interface ToastState {
+  messageList: ReactNode[];
+}
+
+interface ToastAction {
+  removeToast: (node: ReactNode) => void;
+  toast: (node: ReactNode) => void;
+}
+
+interface ToastStore extends ToastState, ToastAction {}
+
+const useToastStore = create<ToastStore>((set, get) => ({
+  messageList: [],
+  removeToast: node => set(state => ({ messageList: state.messageList.filter(msg => msg !== node) })),
+  toast: node =>
+    set(state => {
+      const newMessageList = state.messageList.concat(node);
+      setTimeout(() => {
+        get().removeToast(node);
+      }, 3000);
+      return { messageList: newMessageList };
+    }),
+}));
+
+export const toast = (node: ReactNode) => useToastStore.getState().toast(node);
 
 interface ToastItemProps {
   children: React.ReactNode;
